Switch to crates.io release for wallabag-api 0.3.0
This is the same revision as was previously,
but pulls it from crates.io instead of github.
We need this because the GitHub repository will be deleted very soon,
and also this should depend on stable releases for libraries.

Looks like Cargo.lock was updated to a new version too,
causing the large diff.  Sorry.